+++ src/main/java/com/journal/florist/backend/feature/product/dto/product/ProductMapper.java
@@ -33,7 +33,7 @@ public class ProductMapper {
     private String createdAt;
     private String updatedBy;
     private String updatedAt;
-    private Long weight;
+    private Float weight;
     private String material;
     private BigDecimal materialPrice;
